+++ src/presentation/controllers/signup/signup.spec.ts
@@ -1,8 +1,6 @@
 import { SignUpController } from './signup'
-import { MissingParamError, InvalidParamError, ServerError } from '../errors'
-import { EmailValidador } from '../protocols'
-import { AccountModel } from '../../domain/models/account'
-import { AddAccount, AddAccountModel } from '../../domain/usercases/add-account'
+import { MissingParamError, InvalidParamError, ServerError } from '../../errors'
+import { EmailValidador, AccountModel, AddAccount, AddAccountModel } from './signup-protocols'
 
 const makeEmailValidador = (): EmailValidador => {
   class EmailValidadorStub implements EmailValidador {
