+++ src/presentation/controllers/signup/signup.ts
@@ -1,7 +1,6 @@
-import { HttpResponse, HttpRequest, Controller, EmailValidador } from '../protocols'
-import { MissingParamError, InvalidParamError } from '../errors'
-import { badRequest, serverError } from '../helpers/http-helper'
-import { AddAccount } from '../../domain/usercases/add-account'
+import { HttpResponse, HttpRequest, Controller, EmailValidador, AddAccount } from './signup-protocols'
+import { MissingParamError, InvalidParamError } from '../../errors'
+import { badRequest, serverError } from '../../helpers/http-helper'
 
 export class SignUpController implements Controller {
   private readonly emailValidador: EmailValidador
