+++ src/presentation/protocols/index.ts
@@ -1,3 +1,2 @@
 export * from './controller'
 export * from './http'
-export * from './email-validator'
